% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aft_nl.R
\name{aft_nl}
\alias{aft_nl}
\title{Fit AFT models with non-linear cumulative survival function}
\usage{
aft_nl(
  formula,
  data = list(),
  val = list(),
  outer.args = list(),
  censored = NULL,
  min.lambda = 0.005,
  trace = FALSE,
  ps = TRUE,
  ...
)
}
\arguments{
\item{formula, family, data}{same arguments as in [mgcv::gam].}

\item{...}{further arguments to be passed to [mgcv::gam].}
}
\description{
Fit AFT models with non-linear cumulative survival function
}
