% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff_misi.R
\name{eff_misi}
\alias{eff_misi}
\title{Build single index effect}
\usage{
eff_misi(Xi, basis, a0 = NULL)
}
\arguments{
\item{Xi}{matrix to be projected via single index vector \code{alpha}.}

\item{basis}{function which takes \code{si = Xi\%*\%alpha} as input and returns model
matrix and its derivatives w.r.t. \code{si}.}
}
\description{
Build single index effect
}
