% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_si.R
\name{get.si}
\alias{get.si}
\title{compute single index}
\usage{
get.si(sm, coefs, X = NULL, deriv = 0)
}
\arguments{
\item{X}{a model matrix}

\item{object}{an aft.nl object}
}
\description{
compute single index
}
