% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_afc.R
\name{llk_afc}
\alias{llk_afc}
\title{Log-likelhood of a AFC model}
\usage{
llk_afc(obj, sm, delta, lambda, vpen, mpen)
}
\arguments{
\item{param}{XXX.}

\item{deriv}{XXX.}
}
\description{
XXX.
}
\examples{
library(gamFactory)
n <- 10
y <- rnorm(n)
param <- c(0.5, 1.5) # mu and 1/sigma

# Derivatives of Gaussian log-lik up to order 3
llk_afc(y = y, param = param, deriv = 3)

# Wrap derivatives for compatibility with gamFactory::checkDeriv
obj <- list(
  "d0" = function(param){
    sum(llk_afc(y = y, param = param, deriv = 0)$d0)
  },
  "d1" = function(param){
    colSums(do.call("cbind", llk_afc(y = y, param = param, deriv = 1)$d1))

  },
  "d2" = function(param){
    colSums(do.call("cbind", llk_afc(y = y, param = param, deriv = 2)$d2))

  },
  "d3" = function(param){
    colSums(do.call("cbind", llk_afc(y = y, param = param, deriv = 3)$d3))
  })

check_deriv(obj = obj, param = param, ord = 1:3)
}
