% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHessDrho_htx.R, R/log_htx.R
\name{DHessDrho.htx}
\alias{DHessDrho.htx}
\alias{log.htx}
\title{Log-Hazard function}
\usage{
DHessDrho.htx(o, delta = NULL, llk, DbDr)

\method{log}{htx}(o, deriv = 1, delta = NULL, param = NULL)
}
\description{
Log-Hazard function

Log-Hazard function
}
