% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_aft_nl.R
\name{predict.aft_nl}
\alias{predict.aft_nl}
\title{Prediction from fitted AFT.NL model}
\usage{
\method{predict}{aft_nl}(object, type = "response")
}
\arguments{
\item{object}{a fitted aft.nl object}

\item{type}{When this has value ``response'' the predicted time values are returned.}
}
\description{
Prediction from fitted AFT.NL model
}
