% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_predict_matrix_nested.R
\name{Predict.matrix.nested}
\alias{Predict.matrix.nested}
\title{Predict using nested effects. From gamFactory}
\usage{
\method{Predict.matrix}{nested}(object, data, ...)
}
\description{
Predict using nested effects. From gamFactory
}
