% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_construct_misi_smooth_spec.R
\name{smooth.construct.misi.smooth.spec}
\alias{smooth.construct.misi.smooth.spec}
\title{Monotone increasing single index effects for mgcv}
\usage{
\method{smooth.construct}{misi.smooth.spec}(object, data, knots, unif = FALSE)
}
\description{
Monotone increasing single index effects for mgcv
}
